--- backend/pick-git/src/main/java/com/woowacourse/pickgit/post/application/dto/TokenDto.java
@@ -1,17 +1,0 @@
-package com.woowacourse.pickgit.post.application.dto;
-
-public class TokenDto {
-
-    private String accessToken;
-
-    private TokenDto() {
-    }
-
-    public TokenDto(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-}
