+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/post/application/dto/response/PostResponseDto.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.post.application.dto;
+package com.woowacourse.pickgit.post.application.dto.response;
 
 import java.util.List;
