+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/post/presentation/PostController.java
@@ -5,10 +5,10 @@ import com.woowacourse.pickgit.authentication.domain.user.AppUser;
 import com.woowacourse.pickgit.post.application.CommentRequestDto;
 import com.woowacourse.pickgit.post.application.CommentResponseDto;
 import com.woowacourse.pickgit.post.application.PostService;
-import com.woowacourse.pickgit.post.application.dto.PostRequestDto;
-import com.woowacourse.pickgit.post.application.dto.PostResponseDto;
-import com.woowacourse.pickgit.post.application.dto.RepositoryDto;
-import com.woowacourse.pickgit.post.application.dto.TokenDto;
+import com.woowacourse.pickgit.post.application.dto.request.PostRequestDto;
+import com.woowacourse.pickgit.post.application.dto.response.PostResponseDto;
+import com.woowacourse.pickgit.post.application.dto.response.RepositoryResponseDto;
+import com.woowacourse.pickgit.post.application.dto.request.TokenRequestDto;
 import com.woowacourse.pickgit.post.presentation.dto.PostRequest;
 import java.net.URI;
 import javax.servlet.http.HttpServletRequest;
@@ -82,9 +82,9 @@ public class PostController {
     }
 
     @GetMapping("/github/repositories")
-    public ResponseEntity<RepositoryDto> showRepositories(HttpServletRequest httpServletRequest) {
+    public ResponseEntity<RepositoryResponseDto> showRepositories(HttpServletRequest httpServletRequest) {
         String token = httpServletRequest.getHeader(HttpHeaders.AUTHORIZATION).split(" ")[1];
-        RepositoryDto repositories = postService.showRepositories(new TokenDto(token));
+        RepositoryResponseDto repositories = postService.showRepositories(new TokenRequestDto(token));
 
         return ResponseEntity.ok(repositories);
     }
